pyproject.toml
[project]
name = "ocr-python-gui"
version = "0.1.0"
description = "OCR 圖片文字辨識桌面應用程式"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6",
    "rapidocr-onnxruntime>=1.3",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
